File renaming
the filters now are in a filter directory
and the hash generator is now in the Security file
